feat(backend): move helpers into seperate files ss-182
