+++ apps/backend/src/libs/modules/plugins/auth/libs/helpers/helpers.ts
@@ -1,1 +1,3 @@
 export { checkIsWhiteRoute } from "./check-is-white-route.helper.js";
+export { isTokenValid } from "./is-token-valid.helper.js";
+export { validationAuthHeader } from "./validation-auth-header.helper.js";
